www/requirements.txt
Django==2.2.2
djangorestframework==3.9.4
django_filter==2.1.0
django-cors-headers==3.0.2
psycopg2
